Stocker refactor and staging
