+++ amplify/backend/function/stockerlambda/package.json
@@ -1,0 +1,87 @@
+{
+  "name": "stockerlambda",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "dependencies": {
+    "aws-serverless-express": "^3.3.5",
+    "axios": "^0.24.0",
+    "body-parser": "^1.17.1",
+    "cors": "^2.8.5",
+    "express": "^4.17.3",
+    "luxon": "^3.3.0"
+  },
+  "scripts": {
+    "build": "tsc",
+    "dev": "yarn build && NODE_ENV=dev node src/local.js",
+    "lint": "yarn eslint --ext .js,.ts",
+    "test": "yarn jest --coverage",
+    "test:events": "amplify mock function stockerlambda --event src/event.json"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "node",
+    "collectCoverageFrom": [
+      "src/**/*.ts",
+      "!src/{index,local,app}.ts"
+    ],
+    "coverageThreshold": {
+      "global": {
+        "lines": 99,
+        "branches": 81
+      }
+    }
+  },
+  "eslintConfig": {
+    "extends": [
+      "airbnb-base",
+      "airbnb-typescript/base",
+      "eslint:recommended",
+      "plugin:@typescript-eslint/recommended"
+    ],
+    "rules": {
+      "semi": "error",
+      "no-unexpected-multiline": "error",
+      "arrow-parens": "off",
+      "max-classes-per-file": "off",
+      "@typescript-eslint/lines-between-class-members": "off",
+      "import/prefer-default-export": "off",
+      "@typescript-eslint/ban-ts-comment": "off",
+      "@typescript-eslint/ban-types": "off",
+      "@typescript-eslint/no-use-before-define": "off",
+      "new-cap": "off",
+      "no-underscore-dangle": "off",
+      "import/extensions": "off",
+      "no-param-reassign": "off"
+    },
+    "parserOptions": {
+      "project": "./tsconfig.json"
+    },
+    "ignorePatterns": [
+      "*.js",
+      "*.d.ts"
+    ],
+    "overrides": [
+      {
+        "files": [
+          "src/**/*.ts?(x)",
+          "src/**/*.test.ts?(x)"
+        ]
+      }
+    ]
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@types/aws-serverless-express": "^3.3.7",
+    "@types/cors": "^2.8.15",
+    "@types/express": "^4.17.20",
+    "@types/jest": "^29.5.6",
+    "@types/node": "^20.8.7",
+    "eslint": "^8.6.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "jest": "^27.4.6",
+    "ts-jest": "^29.1.1",
+    "typescript": "^5.2.2"
+  }
+}
