--- amplify/backend/function/stockerlambda/src/errors.js
@@ -1,12 +1,0 @@
-class HTTPError extends Error {
-  constructor(message, status, code) {
-    super(message);
-    this.message = message;
-    this.status = status;
-    this.code = code;
-  }
-}
-
-module.exports = {
-  HTTPError,
-};
