--- amplify/backend/function/stockerlambda/src/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "stockerlambda",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "dependencies": {
-    "aws-serverless-express": "^3.3.5",
-    "axios": "^0.24.0",
-    "luxon": "^3.3.0",
-    "body-parser": "^1.17.1",
-    "cors": "^2.8.5",
-    "express": "^4.17.3"
-  },
-  "jest": {
-    "coverageThreshold": {
-      "global": {
-        "branches": 99
-      }
-    }
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "eslint": "^8.6.0",
-    "jest": "^27.4.6"
-  }
-}
