+++ amplify/backend/function/stockerlambda/ts/app.ts
@@ -1,26 +1,27 @@
-const cors = require('cors');
-const express = require('express');
-const bodyParser = require('body-parser');
-const awsServerlessExpressMiddleware = require('aws-serverless-express/middleware');
-const luxon = require('luxon');
+import cors from 'cors';
+import express from 'express';
+import bodyParser from 'body-parser';
+import awsServerlessExpressMiddleware from 'aws-serverless-express/middleware';
 
-const yahoo = require('./yahoo');
+import * as yahoo  from './yahoo';
+import { Price } from './types';
 
-const app = express()
+export const app = express()
 app.use(bodyParser.json())
 app.use(awsServerlessExpressMiddleware.eventContext())
 
-
-const ALLOWED_ORIGINS = [
-  'http://localhost:3000',
-  'https://app.maffin.io',
-];
-
 app.use(cors({
   origin: (origin, callback) => {
-    if (
-      ALLOWED_ORIGINS.indexOf(origin) !== -1
-      || origin.match(/https:\/\/.*d199ayutgsrlxn\.amplifyapp\.com/g)
+    if (process.env.NODE_ENV === 'dev') {
+      callback(null, true);
+    } else if (
+      process.env.NODE_ENV === 'staging'
+      && origin === 'https://staging.d1w6jie2l5rnr4.amplifyapp.com'
+    ) {
+      callback(null, true);
+    } else if (
+      process.env.NODE_ENV === 'prod'
+      && origin === 'https://app.maffin.io'
     ) {
       callback(null, true);
     } else {
@@ -30,8 +31,8 @@ app.use(cors({
 }));
 
 app.get('/api/prices/live', async (req, res) => {
-  let tickers = req.query.ids;
-  if (!tickers) {
+  let tickers = (req.query.ids as string || '').split(',');
+  if (!tickers.length) {
     res.status(400).json({
       error: 'IDS_REQUIRED',
       description: 'You need to pass \'ids\' queryparam to select specific quotes',
@@ -39,15 +40,13 @@ app.get('/api/prices/live', async (req, res) => {
     return;
   }
 
-  tickers = tickers.split(',');
+  const result: { [ticker: string]: Price } = {};
 
-  const result = {};
-  const promises = [];
-
-  async function callAndSave(ticker) {
-    price = await yahoo.getLiveSummary(ticker);
-    result[ticker] = price;
+  async function callAndSave(ticker: string) {
+    result[ticker] = await yahoo.getPrice(ticker);
   }
+
+  const promises: Promise<void>[] = [];
 
   try {
     tickers.forEach(ticker => promises.push(callAndSave(ticker)));
@@ -66,7 +65,7 @@ app.get('/api/prices/live', async (req, res) => {
 
 app.get('/api/price', async (req, res) => {
   let ticker = req.query.id;
-  let when = Number(req.query.when) || luxon.DateTime.now().toMillis();
+  let when = Number(req.query.when) || undefined;
 
   if (!ticker) {
     res.status(400).json({
@@ -77,7 +76,8 @@ app.get('/api/price', async (req, res) => {
   }
 
   try {
-    price = await yahoo.getPrice(ticker, when);
+    const price = await yahoo.getPrice(ticker as string, when);
+    res.json(price);
   } catch (error) {
     console.log(error);
     res.status(500).json({
@@ -86,8 +86,4 @@ app.get('/api/price', async (req, res) => {
     });
     return;
   }
-
-  res.json(price);
 });
-
-module.exports = app;
