+++ amplify/backend/function/stockerlambda/ts/event.json
