+++ amplify/backend/function/stockerlambda/ts/index.ts
@@ -1,9 +1,9 @@
-const awsServerlessExpress = require('aws-serverless-express');
-const app = require('./app');
+import awsServerlessExpress from 'aws-serverless-express';
+import { app } from './app';
 
 const server = awsServerlessExpress.createServer(app);
 
-exports.handler = (event, context) => {
+exports.handler = (event: any, context: any) => {
   console.log(`EVENT: ${JSON.stringify(event)}`);
   return awsServerlessExpress.proxy(server, event, context, 'PROMISE').promise;
 };
