+++ amplify/backend/function/stockerlambda/ts/local.ts
@@ -1,4 +1,5 @@
-const app = require('./app')
+import { app } from './app';
+
 const port = 3001
 
 app.listen(port)
