+++ amplify/backend/function/stockerlambda/ts/types.d.ts
@@ -1,0 +1,6 @@
+export type Price = {
+  price: number,
+  currency: string,
+  changePct: number,
+  changeAbs: number,
+}
