+++ amplify/backend/function/stockerlambda/tsconfig.json
@@ -1,17 +1,20 @@
 {
   "compilerOptions": {
-    "allowSyntheticDefaultImports": true,
-    "lib": ["dom", "esnext"],
+    "target": "es2016",
     "module": "commonjs",
-    "moduleResolution": "node",
     "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
     "skipLibCheck": true,
-    "resolveJsonModule": true,
+    "rootDir": "./ts",
     "outDir": "./src",
     "baseUrl": "./",
-    "rootDir": "./lib",
     "paths": {
       "src": ["./lib"]
     }
-  }
+  },
+  "exclude": [
+    "node_modules",
+    "ts/__tests__/*"
+  ]
 }
